処理が速くなくても問題なさそうな関数（Ldirect_Trans() と Ldirect_DrawObject() 以外）は高速 RAM から外す
高速 RAM 分が
Code Section Size 8fc=> 408
Data Section Size 4c => 8
Bss Section Size 4 => 0
に減っています。
